requirements.txt
cycler==0.10.0
Django==1.10.4
matplotlib==3.1.0
numpy==1.11.2
Pillow==3.4.2
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
